requirements.txt
debugpy==1.8.9
psycopg2-binary==2.9.9
pytest==8.3.4
pytest-cov==6.0.0
